20-Industry-Use-Cases/23-Energy-Well-Reports/pyproject.toml
[project]
name = "23-energy-well-reports"
version = "0.1.0"
description = "Example using Amazon Bedrock Data Automation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.38.27",
    "pandas>=2.3.1",
    "s3fs==2025.5.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "black>=25.1.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
]
infra = [
    "aws-cdk-lib>=2.205.0",
    "cdk-nag>=2.36.38",
    "constructs>=10.4.2",
]
